小程序整合v3（调整字段，新增推荐模块）

+++ src/main/java/com/example/ode/controller/DishController.java
@@ -38,13 +38,13 @@ public class DishController {
     }
 
     @RequestMapping("/del")
-    public Result delete(@RequestBody List<Integer> ids){
+    public Result delete(@RequestBody List<Long> ids){
         dishService.removeByIds(ids);
         return Result.success();
     }
 
     @GetMapping("/detail/{id}")
-    public Result detail(@PathVariable("id") Integer id){
+    public Result detail(@PathVariable("id") Long id){
         return Result.success(dishService.getById(id));
     }
